Refactor to use env for mock/dev db
