+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:50490/test?","mongoDBName":"test"}
+{"mongoUri":"mongodb://127.0.0.1:50708/test?","mongoDBName":"test"}
