+++ server/server.js
@@ -1,3 +1,4 @@
+process.env.NODE_ENV = 'dev';
 const app = require('./app.js');
 const PORT = 3000;
